maSauvegade

+++ src/polyrallye/ouie/environnement/Sfx.java
@@ -18,8 +18,8 @@ public class Sfx extends Thread {
 	protected float positionY;
 	protected float positionZ;
 
-	public float distance;
-	protected float realDistance;
+	public double distance;
+	protected double realDistance;
 
 	protected boolean isAlive;
 
@@ -90,7 +90,7 @@ public class Sfx extends Thread {
 		vitesse = t;
 	}
 
-	public void setDistance(float d) {
+	public void setDistance(double d) {
 		distance = d;
 	}
